remove screenshot in browser observation
